+++ python/GafferImageUI/ImageStatsUI.py
@@ -39,18 +39,6 @@ import GafferUI
 import GafferImage
 import GafferImageUI
 
-## A function suitable as the postCreator in a NodeMenu.append() call. It
-# sets the region of interest for the node to cover the entire format.
-def postCreate( node, menu ) :
-
-	with node.scriptNode().context() :
-		if node["in"].getInput() :
-			format = node["in"]["format"].getValue()
-		else:
-			format = GafferImage.FormatPlug.getDefaultFormat( node.scriptNode().context() )
-
-	node["area"].setValue( format.getDisplayWindow() )
-
 Gaffer.Metadata.registerNode(
 
 	GafferImage.ImageStats,
@@ -117,6 +105,7 @@ Gaffer.Metadata.registerNode(
 
 			"nodule:type", "",
 			"plugValueWidget:type", "GafferUI.PresetsPlugValueWidget",
+			"userDefault", GafferImage.ImageStats.AreaSource.DisplayWindow,
 
 		],
 
@@ -129,6 +118,7 @@ Gaffer.Metadata.registerNode(
 			""",
 
 			"layout:activator", "areaSourceIsArea",
+			"userDefault", lambda plug : GafferImage.FormatPlug.getDefaultFormat( Gaffer.Context.current() ).getDisplayWindow()
 
 		],